refactor util imports
